third_party/blink/web_tests/wpt_internal/view-transition-on-navigation/at-rule-opt-in-via-cssom.html

<!DOCTYPE html>
<title>View Transitions: @view-transition opt-in programmatically.</title>
<link rel="help" href="https://drafts.csswg.org/css-view-transitions-2/">
<link rel="author" href="mailto:bokan@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
@view-transition {
    navigation: none;
}
</style>
<script>
function optIn() {
    document.styleSheets[0].cssRules[0].navigation = 'auto';
}
const params = new URLSearchParams(location.search);
const is_new_page = params.has('new');

if (!is_new_page) {
  onload = () => requestAnimationFrame(() => requestAnimationFrame(() => {
      optIn();
      location.replace(location.href + '?new');
  }));
} else {
  promise_test(() => {
    optIn();
    return new Promise((resolve) => {
        addEventListener('pagereveal', resolve);
    }).then(event => {
        assert_not_equals(event.viewTransition, null,
            "ViewTransition must be triggered.");
    });
  });
}
</script>
